+++ src/experimental/actions/getCallsReceipt.test.ts
@@ -6,11 +6,11 @@ import { mainnet } from '../../chains/index.js'
 import { createClient } from '../../clients/createClient.js'
 import { custom } from '../../clients/transports/custom.js'
 import { RpcRequestError } from '../../errors/request.js'
-import type { WalletGetCallsStatusReceipt } from '../../types/eip1193.js'
+import type { WalletCallReceipt } from '../../types/eip1193.js'
 import type { Hex } from '../../types/misc.js'
 import { getHttpRpcClient, parseEther } from '../../utils/index.js'
 import { uid } from '../../utils/uid.js'
-import { getCallsStatus } from './getCallsStatus.js'
+import { getCallsReceipt } from './getCallsReceipt.js'
 import { sendCalls } from './sendCalls.js'
 
 type Uid = string
@@ -27,7 +27,7 @@ const getClient = ({
 
         const rpcClient = getHttpRpcClient(localHttpUrl)
 
-        if (method === 'wallet_getCallsStatus') {
+        if (method === 'wallet_getCallsReceipt') {
           const hashes = calls.get(params)
           if (!hashes) return null
           const receipts = await Promise.all(
@@ -52,7 +52,7 @@ const getClient = ({
                 logs: result.logs,
                 status: result.status,
                 transactionHash: result.transactionHash,
-              } satisfies WalletGetCallsStatusReceipt
+              } satisfies WalletCallReceipt
             }),
           )
           return { status: 'CONFIRMED', receipts }
@@ -118,7 +118,7 @@ test('default', async () => {
 
   await mine(testClient, { blocks: 1 })
 
-  const { status, receipts } = await getCallsStatus(client, { id })
+  const { status, receipts } = await getCallsReceipt(client, { id })
   expect(status).toMatchInlineSnapshot(`"CONFIRMED"`)
   expect(receipts![0].blockHash).toBeDefined()
   expect(