+++ src/experimental/actions/getCallsReceipt.ts
@@ -3,50 +3,50 @@ import type { Transport } from '../../clients/transports/createTransport.js'
 import type { ErrorType } from '../../errors/utils.js'
 import type { Account } from '../../types/account.js'
 import type { Chain } from '../../types/chain.js'
-import type { WalletGetCallsStatusReturnType } from '../../types/eip1193.js'
+import type { WalletGetCallsReceiptReturnType } from '../../types/eip1193.js'
 import type { Prettify } from '../../types/utils.js'
 import type { RequestErrorType } from '../../utils/buildRequest.js'
 import { hexToBigInt } from '../../utils/encoding/fromHex.js'
 import { receiptStatuses } from '../../utils/formatters/transactionReceipt.js'
 
-export type GetCallsStatusParameters = { id: string }
+export type GetCallsReceiptParameters = { id: string }
 
-export type GetCallsStatusReturnType = Prettify<
-  WalletGetCallsStatusReturnType<bigint, 'success' | 'reverted'>
+export type GetCallsReceiptReturnType = Prettify<
+  WalletGetCallsReceiptReturnType<bigint, 'success' | 'reverted'>
 >
 
-export type GetCallsStatusErrorType = RequestErrorType | ErrorType
+export type GetCallsReceiptErrorType = RequestErrorType | ErrorType
 
 /**
  * Returns the status of a call batch that was sent via `sendCalls`.
  *
- * - Docs: https://viem.sh/experimental/actions/getCallsStatus
- * - JSON-RPC Methods: [`wallet_getCallsStatus`](https://eips.ethereum.org/EIPS/eip-5792)
+ * - Docs: https://viem.sh/experimental/actions/getCallsReceipt
+ * - JSON-RPC Methods: [`wallet_getCallsReceipt`](https://eips.ethereum.org/EIPS/eip-5792)
  *
  * @param client - Client to use
- * @returns Status of the calls. {@link GetCallsStatusReturnType}
+ * @returns Status of the calls. {@link GetCallsReceiptReturnType}
  *
  * @example
  * import { createWalletClient, custom } from 'viem'
  * import { mainnet } from 'viem/chains'
- * import { getCallsStatus } from 'viem/wallet'
+ * import { getCallsReceipt } from 'viem/wallet'
  *
  * const client = createWalletClient({
  *   chain: mainnet,
  *   transport: custom(window.ethereum),
  * })
- * const { receipts, status } = await getCallsStatus(client, { id: '0xdeadbeef' })
+ * const { receipts, status } = await getCallsReceipt(client, { id: '0xdeadbeef' })
  */
-export async function getCallsStatus<
+export async function getCallsReceipt<
   chain extends Chain | undefined,
   account extends Account | undefined = undefined,
 >(
   client: Client<Transport, chain, account>,
-  parameters: GetCallsStatusParameters,
-): Promise<GetCallsStatusReturnType> {
+  parameters: GetCallsReceiptParameters,
+): Promise<GetCallsReceiptReturnType> {
   const { id } = parameters
   const { receipts, status } = await client.request({
-    method: 'wallet_getCallsStatus',
+    method: 'wallet_getCallsReceipt',
     params: id,
   })
   return {